Refactored variable code

+++ SlgToFit/SlfEntities/GeneralInformation.cs
@@ -9,14 +9,14 @@ namespace SlfToFit.SlfEntities
 		string sport,
 		string gUID,
 		int age,
-		int altitudeDifferencesDownhill,
-		int altitudeDifferencesUphill,
+		ushort altitudeDifferencesDownhill,
+		ushort altitudeDifferencesUphill,
 		int averageAltitude,
-		int averageCadence,
-		int averageHeartrate,
+		byte averageCadence,
+		byte averageHeartrate,
 		float averageInclineUphill,
 		float averageInclineDownhill,
-		int averagePower,
+		ushort averagePower,
 		float averatePowerWattPerKG,
 		float averageRiseRate,
 		float averageRiseRateUphill,
@@ -30,7 +30,7 @@ namespace SlfToFit.SlfEntities
 		int bodyHeight,
 		int bodyWeight,
 		bool calibration,
-		int calories,
+		ushort calories,
 		string dataType,
 		string description,
 		float distance,
@@ -51,19 +51,19 @@ namespace SlfToFit.SlfEntities
 		float longitudeEnd,
 		float longitudeStart,
 		float manualTemperature,
-		int maximumAltitude,
-		int maximumCadence,
-		int maximumHeartrate,
+		float maximumAltitude,
+		byte maximumCadence,
+		byte maximumHeartrate,
 		float maximumIncline,
 		float maximumInclineDownhill,
 		float maximumInclineUphill,
-		int maximumPower,
+		ushort maximumPower,
 		float maximumRiseRate,
 		float maximumSpeed,
 		float maximumTemperature,
-		int minimumAltitude,
+		float minimumAltitude,
 		int minimumCadence,
-		int minimumHeartrate,
+		byte minimumHeartrate,
 		float minimumIncline,
 		int minimumPower,
 		float minimumRiseRate,
@@ -131,17 +131,17 @@ namespace SlfToFit.SlfEntities
 	{
 		public readonly User User = user;                                                   // User that recorded the activity
 		public readonly Sport Sport = SportMap(sport);                                      // Type of sport that is recorded
-		public readonly SubSport SubSport = SubSportMap(sport);							// Type of subsport that is recorded
+		public readonly SubSport SubSport = SubSportMap(sport);								// Type of subsport that is recorded
 		public readonly string GUID = gUID;                                                 // GUID of the activity
 		public readonly int Age = age;                                                      // Age of the user
-		public readonly int AltitudeDifferencesDownhill = altitudeDifferencesDownhill;      // Differences in altitude downhill
-		public readonly int AltitudeDifferencesUphill = altitudeDifferencesUphill;          // Differences in altitude uphill
+		public readonly ushort AltitudeDifferencesDownhill = altitudeDifferencesDownhill;   // Differences in altitude downhill
+		public readonly ushort AltitudeDifferencesUphill = altitudeDifferencesUphill;       // Differences in altitude uphill
 		public readonly int AverageAltitude = averageAltitude;                              // Average altitude recorded
-		public readonly int AverageCadence = averageCadence;                                // Average cadence
-		public readonly int AverageHeartrate = averageHeartrate;                            // Average heartrate
+		public readonly byte AverageCadence = averageCadence;                               // Average cadence
+		public readonly byte AverageHeartrate = averageHeartrate;                           // Average heartrate
 		public readonly float AverageInclineUphill = averageInclineUphill;                  // Average incline uphill
 		public readonly float AverageInclineDownhill = averageInclineDownhill;              // Average incline downhill
-		public readonly int AveragePower = averagePower;                                    // Average power
+		public readonly ushort AveragePower = averagePower;                                 // Average power
 		public readonly float AveratePowerWattPerKG = averatePowerWattPerKG;                // Average power per kilogram
 		public readonly float AverageRiseRate = averageRiseRate;                            // Average rise rate
 		public readonly float AverageRiseRateUphill = averageRiseRateUphill;                // Average rise rate uphill
@@ -149,13 +149,13 @@ namespace SlfToFit.SlfEntities
 		public readonly float AverageSpeed = averageSpeed;                                  // Average speed
 		public readonly float AverageSpeedDownhill = averageSpeedDownhill;                  // Average speed downhill
 		public readonly float AverageSpeedUphill = averageSpeedUphill;                      // Average speed uphill
-		public readonly float AverageTemperature = averageTemperature;                      // Average temperature
+		public readonly sbyte AverageTemperature = (sbyte)averageTemperature;               // Average temperature
 		public readonly string Bike = bike;                                                 // Bike used for the activity
 		public readonly int BikeWeight = bikeWeight;                                        // Weight of the bike
 		public readonly int BodyHeight = bodyHeight;                                        // Height of the user
 		public readonly int BodyWeight = bodyWeight;                                        // Weight of the user
 		public readonly bool Calibration = calibration;                                     // Indicates if calibration was performed
-		public readonly int Calories = calories;                                            // Calories burned
+		public readonly ushort Calories = calories;                                            // Calories burned
 		public readonly string DataType = dataType;                                         // Type of data
 		public readonly string Description = description;                                   // Description of the activity
 		public readonly float Distance = distance;                                          // Total distance covered
@@ -176,26 +176,26 @@ namespace SlfToFit.SlfEntities
 		public readonly float LongitudeEnd = longitudeEnd;                                  // End longitude
 		public readonly float LongitudeStart = longitudeStart;                              // Start longitude
 		public readonly float ManualTemperature = manualTemperature;                        // Manually recorded temperature
-		public readonly int MaximumAltitude = maximumAltitude;                              // Maximum altitude
-		public readonly int MaximumCadence = maximumCadence;                                // Maximum cadence
-		public readonly int MaximumHeartrate = maximumHeartrate;                            // Maximum heart rate
+		public readonly float MaximumAltitude = maximumAltitude;                            // Maximum altitude
+		public readonly byte MaximumCadence = maximumCadence;                               // Maximum cadence
+		public readonly byte MaximumHeartrate = maximumHeartrate;                           // Maximum heart rate
 		public readonly float MaximumIncline = maximumIncline;                              // Maximum incline
 		public readonly float MaximumInclineDownhill = maximumInclineDownhill;              // Maximum incline downhill
 		public readonly float MaximumInclineUphill = maximumInclineUphill;                  // Maximum incline uphill
-		public readonly int MaximumPower = maximumPower;                                    // Maximum power
+		public readonly ushort MaximumPower = maximumPower;                                 // Maximum power
 		public readonly float MaximumRiseRate = maximumRiseRate;                            // Maximum rise rate
 		public readonly float MaximumSpeed = maximumSpeed;                                  // Maximum speed
-		public readonly float MaximumTemperature = maximumTemperature;                      // Maximum temperature
-		public readonly int MinimumAltitude = minimumAltitude;                              // Minimum altitude
+		public readonly sbyte MaximumTemperature = (sbyte)maximumTemperature;               // Maximum temperature
+		public readonly float MinimumAltitude = minimumAltitude;                            // Minimum altitude
 		public readonly int MinimumCadence = minimumCadence;                                // Minimum cadence
 		public readonly int MinimumHeartrate = minimumHeartrate;                            // Minimum heart rate
 		public readonly float MinimumIncline = minimumIncline;                              // Minimum incline
 		public readonly int MinimumPower = minimumPower;                                    // Minimum power
 		public readonly float MinimumRiseRate = minimumRiseRate;                            // Minimum rise rate
 		public readonly float MinimumSpeed = minimumSpeed;                                  // Minimum speed
-		public readonly float MinimumTemperature = minimumTemperature;                      // Minimum temperature
+		public readonly sbyte MinimumTemperature = (sbyte)minimumTemperature;               // Minimum temperature
 		public readonly string Name = name;                                                 // Name of the activity
-		public readonly float PauseTime = pauseTime/100f;                                    // Total pause time
+		public readonly float PauseTime = pauseTime/100f;                                   // Total pause time
 		public readonly int PowerZone1Start = powerZone1Start;                              // Start of power zone 1
 		public readonly int PowerZone2Start = powerZone2Start;                              // Start of power zone 2
 		public readonly int PowerZone3Start = powerZone3Start;                              // Start of power zone 3
@@ -206,43 +206,43 @@ namespace SlfToFit.SlfEntities
 		public readonly int PowerZone7End = powerZone7End;                                  // End of power zone 7
 		public readonly int Rating = rating;                                                // Rating of the activity
 		public readonly int Feeling = feeling;                                              // User's feeling during the activity
-		public readonly float TrainingTimeDownhill = trainingTimeDownhill/100f;              // Training time downhill
-		public readonly float TrainingTimeUphill = trainingTimeUphill/100f;                  // Training time uphill
+		public readonly float TrainingTimeDownhill = trainingTimeDownhill/100f;             // Training time downhill
+		public readonly float TrainingTimeUphill = trainingTimeUphill/100f;                 // Training time uphill
 		public readonly int SamplingRate = samplingRate;                                    // Sampling rate
 		public readonly int ShoulderWidth = shoulderWidth;                                  // Shoulder width of the user
 		public readonly System.DateTime StartDate = startDate;                              // Start date of the activity
 		public readonly bool Statistic = statistic;                                         // Indicates if it's a statistical activity
-		public readonly float TimeInIntensityZone1 = timeInIntensityZone1/100f;              // Time in intensity zone 1
-		public readonly float TimeInIntensityZone2 = timeInIntensityZone2/100f;              // Time in intensity zone 2
-		public readonly float TimeInIntensityZone3 = timeInIntensityZone3/100f;              // Time in intensity zone 3
-		public readonly float TimeInIntensityZone4 = timeInIntensityZone4/100f;              // Time in intensity zone 4
-		public readonly float TimeInPowerZone1 = timeInPowerZone1/100f;                      // Time in power zone 1
-		public readonly float TimeInPowerZone2 = timeInPowerZone2/100f;                      // Time in power zone 2
-		public readonly float TimeInPowerZone3 = timeInPowerZone3/100f;                      // Time in power zone 3
-		public readonly float TimeInPowerZone4 = timeInPowerZone4/100f;                      // Time in power zone 4
-		public readonly float TimeInPowerZone5 = timeInPowerZone5/100f;                      // Time in power zone 5
-		public readonly float TimeInPowerZone6 = timeInPowerZone6/100f;                      // Time in power zone 6
-		public readonly float TimeInPowerZone7 = timeInPowerZone7/100f;                      // Time in power zone 7
-		public readonly float TimeOverIntensityZone = timeOverIntensityZone/100f;            // Time over intensity zone
-		public readonly float TimeUnderIntensityZone = timeUnderIntensityZone/100f;          // Time under intensity zone
+		public readonly float TimeInIntensityZone1 = timeInIntensityZone1/100f;             // Time in intensity zone 1
+		public readonly float TimeInIntensityZone2 = timeInIntensityZone2/100f;             // Time in intensity zone 2
+		public readonly float TimeInIntensityZone3 = timeInIntensityZone3/100f;             // Time in intensity zone 3
+		public readonly float TimeInIntensityZone4 = timeInIntensityZone4/100f;             // Time in intensity zone 4
+		public readonly float TimeInPowerZone1 = timeInPowerZone1/100f;                     // Time in power zone 1
+		public readonly float TimeInPowerZone2 = timeInPowerZone2/100f;                     // Time in power zone 2
+		public readonly float TimeInPowerZone3 = timeInPowerZone3/100f;                     // Time in power zone 3
+		public readonly float TimeInPowerZone4 = timeInPowerZone4/100f;                     // Time in power zone 4
+		public readonly float TimeInPowerZone5 = timeInPowerZone5/100f;                     // Time in power zone 5
+		public readonly float TimeInPowerZone6 = timeInPowerZone6/100f;                     // Time in power zone 6
+		public readonly float TimeInPowerZone7 = timeInPowerZone7/100f;                     // Time in power zone 7
+		public readonly float TimeOverIntensityZone = timeOverIntensityZone/100f;           // Time over intensity zone
+		public readonly float TimeUnderIntensityZone = timeUnderIntensityZone/100f;         // Time under intensity zone
 		public readonly int TrackProfile = trackProfile;                                    // Profile of the track
-		public readonly float TrainingTime = trainingTime/100f;                              // Total training time
+		public readonly float TrainingTime = trainingTime/100f;                             // Total training time
 		public readonly int UnitId = unitId;                                                // Unit ID
 		public readonly int Weather = weather;                                              // Weather conditions
 		public readonly int WheelSize = wheelSize;                                          // Size of the wheel
 		public readonly int Wind = wind;                                                    // Wind conditions
-		public readonly uint WorkinJ = workinKJ*1000;                                        // Work in kilojoules
-		public readonly float Best5KTime = best5KTime/100f;                                  // Best 5K time
+		public readonly uint WorkinJ = workinKJ*1000;                                       // Work in kilojoules
+		public readonly float Best5KTime = best5KTime/100f;                                 // Best 5K time
 		public readonly float Best5KEntry = best5KEntry;                                    // Entry for best 5K time
 		public readonly float Best20minPower = best20minPower;                              // Best 20-minute power
 		public readonly float Best20minPowerEntry = best20minPowerEntry;                    // Entry for best 20-minute power
-		public readonly float PowerNP = powerNP;                                            // Normalized power
+		public readonly ushort PowerNP = (ushort)powerNP;                                   // Normalized power
 		public readonly float PowerTSS = powerTSS;                                          // Training stress score
-		public readonly float PowerFTP = powerFTP;                                          // Functional threshold power
+		public readonly ushort PowerFTP = (ushort)powerFTP;                                 // Functional threshold power
 		public readonly float PedalingTime = pedalingTime/100f;								// Total pedaling time
 		public readonly float PedalingIndex = pedalingIndex;								// Pedaling index
-		public readonly float AverageBalanceRight = averageBalanceRight;                    // Average balance - right
-		public readonly float AverageBalanceLeft = averageBalanceLeft;                      // Average balance - left
+		public readonly ushort AverageBalanceRight = (ushort)averageBalanceRight;           // Average balance - right
+		public readonly ushort AverageBalanceLeft = (ushort)averageBalanceLeft;             // Average balance - left
 		public readonly float PowerIF = powerIF;                                            // Intensity factor
 		public readonly float TorqueEffectLeft = torqueEffectLeft;                          // Torque effect - left
 		public readonly float TorqueEffectRight = torqueEffectRight;                        // Torque effect - right
